+++ include/roq/format_str.h
@@ -4,8 +4,8 @@
 
 #include <fmt/format.h>
 
+#include <array>
 #include <string_view>
-#include <tuple>
 
 #include "roq/source_location.h"
 
@@ -14,6 +14,39 @@
 // strongly inspired by the discussion here: https://github.com/fmtlib/fmt/issues/2022
 
 namespace roq {
+
+namespace detail {
+// may truncate to N
+template <std::size_t N>
+struct static_string final {
+#if __cplusplus >= 202002L
+  consteval static_string(const std::string_view &sv)
+#else
+  constexpr static_string(const std::string_view &sv)
+#endif
+      : length_(std::min(N, std::size(sv))), buffer_(create(sv, length_)) {
+  }
+
+  static_string(const static_string &) = delete;
+  static_string(static_string &&) = delete;
+
+  constexpr operator std::string_view() const { return {std::data(buffer_), length_}; }
+
+ protected:
+  static constexpr auto create(const std::string_view &sv, std::size_t length) {
+    std::array<char, N> buffer;
+    for (std::size_t i = 0; i < length; ++i)
+      buffer[i] = sv[i];
+    for (std::size_t i = length; i < N; ++i)
+      buffer[i] = 0xEF;  // debug
+    return buffer;
+  }
+
+ private:
+  const std::size_t length_;
+  const std::array<char, N> buffer_;
+};
+}  // namespace detail
 #if __cplusplus >= 202002L
 template <typename... Args>
 struct basic_format_str final {
@@ -28,7 +61,7 @@ struct basic_format_str final {
   }
 
   const fmt::string_view str_;
-  const std::string_view file_name_;
+  const detail::static_string<32> file_name_;
   const std::uint32_t line_;
 
  private:
@@ -47,11 +80,12 @@ struct basic_format_str final {
 #else
 template <typename... Args>
 struct basic_format_str {
-  basic_format_str(const std::string_view &str, const source_location &loc = source_location::current())  // NOLINT
+  constexpr basic_format_str(
+      const std::string_view &str, const source_location &loc = source_location::current())  // NOLINT
       : str_(str), file_name_(extract_basename(loc.file_name())), line_(loc.line()) {}
 
   const fmt::string_view str_;
-  const std::string_view file_name_;
+  const detail::static_string<32> file_name_;
   const std::uint32_t line_;
 
  private:
@@ -73,3 +107,16 @@ template <typename... Args>
 using format_str = basic_format_str<fmt::type_identity_t<Args>...>;
 
 }  // namespace roq
+
+template <size_t N>
+struct fmt::formatter<roq::detail::static_string<N> > {
+  template <typename Context>
+  constexpr auto parse(Context &ctx) {
+    return std::begin(ctx);
+  }
+  template <typename Context>
+  auto format(const roq::detail::static_string<N> &value, Context &ctx) {
+    using namespace std::literals;
+    return fmt::format_to(ctx.out(), "{}"sv, static_cast<std::string_view>(value));
+  }
+};